re-organizes stuff
