+++ UnityProject/Assets/Scripts/Core/Lifecycle/PlayerSpawn.cs
@@ -414,7 +414,7 @@ public static class PlayerSpawn
 			var dummy = ServerCreatePlayer(spawnTransform.position.RoundToInt());
 
 			CharacterSettings randomSettings = new CharacterSettings();
-			randomSettings = CharacterCustomization.RandomizeCharacterSettings();
+			randomSettings = CharacterSettings.RandomizeCharacterSettings();
 
 			ServerTransferPlayer(null, dummy, null, Event.PlayerSpawned, randomSettings);
