+++ UnityProject/Assets/Scripts/UI/Systems/Lobby/CharacterSettings.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using UI.CharacterCreator;
+using UnityEngine;
 
 /// <summary>
 /// Class containing all character preferences for a player
@@ -246,4 +247,117 @@ public class CharacterSettings
 				return "have";
 		}
 	}
+
+	#region StaticCustomizationFunctions
+
+	public static CharacterSettings RandomizeCharacterSettings()
+	{
+		CharacterSettings random = new CharacterSettings();
+		// Randomise gender
+		Type gender = typeof(BodyType);
+		Array genders = gender.GetEnumValues();
+		int index = UnityEngine.Random.Range(0,3);
+		random.BodyType = (BodyType)genders.GetValue(index);
+
+		//Randomises player name and age and skin tone.
+		random.Name = RandomizeCharacterName(random);
+		random.Age  = UnityEngine.Random.Range(19, 78);
+		random.SkinTone = RandomizeCharacterSkinToneHTMLString(random);
+		random.SerialisedExternalCustom = GetRandomUnderwearCustomisation(random);
+
+		//Randomises player accents. (Italian, Scottish, etc)
+		random.Speech = RandomizeCharachterAccent();
+
+		return random;
+	}
+
+	public static List<ExternalCustomisation> GetRandomUnderwearCustomisation(CharacterSettings data)
+	{
+		var RaceData = GetRaceData(data);
+		List<ExternalCustomisation> externalCustomisations = new List<ExternalCustomisation>();
+
+		void logic(CustomisationAllowedSetting setting)
+		{
+			PlayerCustomisationData customizationToAdd = setting.CustomisationGroup.PlayerCustomisations.PickRandom();
+			ExternalCustomisation newExternalCustomisation = new ExternalCustomisation();
+			newExternalCustomisation.Key = customizationToAdd.name;
+			newExternalCustomisation.SerialisedValue = SerialiseCustomizationData(customizationToAdd);
+			externalCustomisations.Add(newExternalCustomisation);
+		}
+
+		foreach (CustomisationAllowedSetting customisation in RaceData.Base.CustomisationSettings)
+		{
+			if (customisation.CustomisationGroup.name == "PlayerUnderShirt") logic(customisation);
+			if (customisation.CustomisationGroup.name == "PlayerUnderWear") logic(customisation);
+			if (customisation.CustomisationGroup.name == "PlayerSocks") logic(customisation);
+		}
+
+		return externalCustomisations;
+	}
+
+	public static CustomisationClass SerialiseCustomizationData(PlayerCustomisationData data)
+	{
+		var newcurrentSetting = new CustomisationClass();
+		newcurrentSetting.Colour = "#" + ColorUtility.ToHtmlStringRGB(new Color(UnityEngine.Random.Range(0.1f, 1f),
+			UnityEngine.Random.Range(0.1f, 1f), UnityEngine.Random.Range(0.1f, 1f), 1f));
+		newcurrentSetting.SelectedName = data.Name;
+		return newcurrentSetting;
+	}
+
+	public static string RandomizeCharacterSkinToneHTMLString(CharacterSettings data)
+	{
+		var RaceData = GetRaceData(data);
+
+		List<Color> raceSkinColors = RaceData.Base.SkinColours;
+		if (raceSkinColors.Count != 0)
+		{
+			return "#" +
+			       ColorUtility.ToHtmlStringRGB(raceSkinColors[UnityEngine.Random.Range(0, raceSkinColors.Count - 1)]);
+		}
+
+		return "#" + ColorUtility.ToHtmlStringRGBA(new Color(UnityEngine.Random.Range(0.1f, 1f),
+			UnityEngine.Random.Range(0.1f, 1f),
+			UnityEngine.Random.Range(0.1f, 1f), 1f));
+	}
+
+	public static string RandomizeCharacterName(CharacterSettings data, bool isNotPlayerDebug = true)
+	{
+		switch (data.BodyType)
+		{
+			case BodyType.Male:
+				return StringManager.GetRandomMaleName();
+			case BodyType.Female:
+				return StringManager.GetRandomFemaleName();
+			default:
+				if (isNotPlayerDebug) return StringManager.GetRandomName(Gender.NonBinary);
+				else return "Cuban Pete";
+		}
+	}
+
+	public static PlayerHealthData GetRaceData(CharacterSettings data)
+	{
+		foreach (var Race in RaceSOSingleton.Instance.Races)
+		{
+			if (Race.name == data.Species) return Race;
+		}
+
+		Logger.LogError($"Unable to get PlayerHealthData from {data.Name} / {data.Username}");
+		return new PlayerHealthData();
+	}
+
+	public static Speech RandomizeCharachterAccent()
+	{
+		int accentChance = UnityEngine.Random.Range(0, 100);
+		if (accentChance <= 35)
+		{
+			Type accent = typeof(Speech);
+			Array accents = accent.GetEnumValues();
+			int index = UnityEngine.Random.Range(0, 7);
+			return (Speech) accents.GetValue(index);
+		}
+
+		return Speech.None;
+	}
+
+	#endregion
 }
